QueryFrequencia cardiaca resolvida

+++ BeActiv/WorkoutView.swift
@@ -13,7 +13,7 @@ struct Workout: Identifiable{
     let duration: Int
     let calories: Int
     let distance: Int
-    let frequencyHeart: Int
+    let frequencyHeart: Double
 }
 
 struct WorkoutView: View {
